% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_inla.R
\docType{methods}
\name{my_inla}
\alias{my_inla}
\alias{my_inla,ANY,ANY-method}
\alias{my_inla,my_inla,inla-method}
\alias{my_inla,my_inla-methods}
\title{Create a my_inla object}
\usage{
my_inla(data, ..., model.fit)

\S4method{my_inla}{ANY,ANY}(data, ..., model.fit)

\S4method{my_inla}{my_inla,inla}(data, ..., model.fit)
}
\arguments{
\item{data}{the dataset}

\item{...}{other arguments}

\item{model.fit}{The fitted model}
}
\description{
Create a my_inla object
}

