% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_lmer.R, R/my_lmer_model_class.R
\docType{methods}
\name{my_lmer}
\alias{my_lmer}
\alias{my_lmer,ANY,ANY,ANY,ANY,ANY-method}
\alias{my_lmer,lmerMod-methods}
\alias{my_lmer,my_lmer,ANY,ANY,ANY,lmerMod-method}
\alias{my_lmer,my_lmer-methods}
\title{Create a my_lmer object}
\usage{
my_lmer(data, meta1, status, weight, model.fit)

\S4method{my_lmer}{ANY,ANY,ANY,ANY,ANY}(data, meta1, status, weight, model.fit)

\S4method{my_lmer}{my_lmer,ANY,ANY,ANY,lmerMod}(data, meta1, status, weight,
  model.fit)
}
\arguments{
\item{data}{the dataset}

\item{meta1}{the first metadata}

\item{status}{the status of the analysis}

\item{weight}{The optional weight}

\item{model.fit}{The fitted model}
}
\description{
Create a my_lmer object
}

