% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_lmer_model_class.R
\docType{class}
\name{my_lmer_model-class}
\alias{my_lmer_model-class}
\title{The my_lmer_model class}
\description{
The my_lmer_model class
}
\section{Slots}{

  \describe{
   \item{\code{Model}}{a lmer model}
  }
}

