% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_virtual_class.R
\docType{class}
\name{my_virtual-class}
\alias{my_virtual-class}
\title{The virtual class}
\description{
The virtual class
}
\section{Slots}{

  \describe{
   \item{\code{Data}}{a data.frame}
   \item{\code{Meta1}}{The first meta data slot}
   \item{\code{Meta2}}{The second meta data slot}
   \item{\code{Meta3}}{The third meta data slot}
   \item{\code{Meta4}}{The fourth meta data slot}
   \item{\code{Meta5}}{The fifth meta data slot}
   \item{\code{Meta6}}{The sixth meta data slot}
   \item{\code{Status}}{a description of the status}
  }
}

