% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_lmer_class.R
\docType{class}
\name{my_lmer-class}
\alias{my_lmer-class}
\title{The my_lmer class}
\description{
The my_lmer class
}
\section{Slots}{

  \describe{
   \item{\code{Weight}}{the name of the weight variable or "" if unweighted}
   \item{\code{Model}}{Either a lmer model or NULL}
  }
}

