% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_lmer.R
\docType{methods}
\name{my_lmer}
\alias{my_lmer}
\alias{my_lmer,ANY,ANY-method}
\alias{my_lmer,my_lmer,lmerMod-method}
\alias{my_lmer,my_lmer-methods}
\title{Create a my_lmer object}
\usage{
my_lmer(data, ..., model.fit)

\S4method{my_lmer}{ANY,ANY}(data, ..., model.fit)

\S4method{my_lmer}{my_lmer,lmerMod}(data, ..., model.fit)
}
\arguments{
\item{data}{the dataset}

\item{...}{other arguments}

\item{model.fit}{The fitted model}
}
\description{
Create a my_lmer object
}

