% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_virtual_class.R
\docType{class}
\name{my_virtual-class}
\alias{my_virtual-class}
\title{The virtual class}
\description{
The virtual class
}
\section{Slots}{

  \describe{
   \item{\code{Data}}{a data.frame}
   \item{\code{Meta}}{The meta data slot}
   \item{\code{Status}}{a description of the status}
  }
}

